SendReport function for sending dynamically generated content in body of email
I added a new function in PythonEvents.cs to be called from a python
script such as MorningBatch.
EmailReport is designed to be very similar to EmailContent, except that
the body of the email is generated by a python script instead of being
pulled from an static file.
I also added a support function in QueryFunction.cs.  QueryList is
designed to run a pre-saved query referenced by name which is passed in
as a string in the function call.
For each people record returned from the query, we save a series of key
values into a list (PersonList).
The individual lists are combined into a combined list and returned from
the function.
These values can then be utilized in the Python Script to generate
custom reports.