Overloaded EmailReport
I added an Overloaded version of EmailReport which adds variables in the
function call for QueryName and QueryDescription. The original version
of EmailReport
required you to embed the query name in the Python Script.  This version
of the function permits you to have a generic Python script
and then call it multiple times with a different query and description
each time.